Split the functionality in two apps, loans and frontend. The new frontend applications will serve the front end, while the loans will be a rest framework based app to provide the backend functionality to the frontend
