+++ growthstreet/urls.py
@@ -1,7 +1,8 @@
 from django.conf.urls import include, url
 from django.contrib import admin
-#Import all views from the loans applications, as we are using all of them.
+#Import all views from the loans and frontend applications, as we are using all of them.
 from loans.views import *
+from frontend.views import *
 
 urlpatterns = [
     url(r'^admin/', include(admin.site.urls)),
